Handle link coloring in pre elements
Some mobile browsers can interpret numbers as phone numbers and create links automatically, this will color those links so they are visible. In case of blue pre element links were also blue, this will force links to be same color as text. This visual bug is reproducable on "Puma" we browser for iOS.